macosx: updated used Cocoa methods to the current definitions and fixed a few logic errors, which only occur in 64 bit mode

+++ modules/gui/macosx/bookmarks.m
@@ -365,7 +365,7 @@ static VLCBookmarks *_o_sharedInstance = nil;
  * data source methods
  *****************************************************************************/
 
-- (int)numberOfRowsInTableView:(NSTableView *)theDataTable
+- (NSInteger)numberOfRowsInTableView:(NSTableView *)theDataTable
 {
     /* return the number of bookmarks */
     input_thread_t * p_input = pl_CurrentInput( VLCIntf );
@@ -386,7 +386,7 @@ static VLCBookmarks *_o_sharedInstance = nil;
 }
 
 - (id)tableView:(NSTableView *)theDataTable objectValueForTableColumn:
-    (NSTableColumn *)theTableColumn row: (int)row
+    (NSTableColumn *)theTableColumn row: (NSInteger)row
 {
     /* return the corresponding data as NSString */
     input_thread_t * p_input = pl_CurrentInput( VLCIntf );
